Updated Product Weight to Compatible Data Type

+++ src/main/java/com/teamtiger/productservice/products/models/GetProductDTO.java
@@ -6,7 +6,7 @@ public record GetProductDTO(
      UUID id,
      String name,
      double retailPrice,
-     int weight
+     double weight
 )
 {
 }
